Provide new feature add product

+++ MilkTea/src/main/java/PTTKYC/MilkTea/Repository/SanPhamRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SanPhamRepository extends PagingAndSortingRepository< SanPham,Long >{
+public interface SanPhamRepository extends PagingAndSortingRepository< SanPham,Integer >{
 }
